Fix the user regex
For explanation:  There was an erronous `|` in the id matching portion.
Slack sends users in the format of `<@U1234|Bob>` where the pipe delimits
the "human readable fallback" portion.  But users can also be sent as `<@U1234>`

The old regex worked fine with both of these situations, but the extra `|` in the first part of the match should not be there.